Renewed \maketitle into a new title page style
The title is now based on the Titlepages by Apple. Extra authors can be
included by doing  <first> \\ <second> .... in the \author.
The \cover is by default empty, but can be enabled, by renewing the \cover.
There is a \subtitle for adding a small subtitle in grey below the orignal title