Changed Fonts to DnD style fonts. Use XeLaTeX
XeLaTex must now be used to compile the document.
The DnDLolth font is used for chapters and sections.
The Mentor Sans Std font is used overall in the document.

Probably should be looking into Mentor Std as a possible font
and now and then be able to switch font from the document.